% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{refs_tiles}
\alias{refs_tiles}
\title{references as tiles}
\usage{
refs_tiles(x)
}
\arguments{
\item{x}{file path or url}
}
\value{
dataframe with xmin,xmax,ymin,ymax add to the tiffs_refs output
}
\description{
Add extent information xmin,xmax,ymin,ymax to refs table
}
\examples{
vfile <- system.file("extdata/volcano-cog.tif", package = "rustycogs", mustWork = TRUE)
refs <- refs_tiles(vfile)
dim(refs)
str(refs)
}
\seealso{
tiles_to_matrix
}
