% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile_to_array}
\alias{tile_to_array}
\title{Convert tile data to a matrix or array}
\usage{
tile_to_array(tile)
}
\arguments{
\item{tile}{A tile result from \code{\link[=tiff_tile]{tiff_tile()}}.}
}
\value{
A matrix (single band) or 3D array (multi-band).
}
\description{
Convert tile data to a matrix or array
}
