% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{tiles_to_matrix}
\alias{tiles_to_matrix}
\title{Turn already-read data into a full matrix (from tiff_read)}
\usage{
tiles_to_matrix(refs)
}
\description{
Turn already-read data into a full matrix (from tiff_read)
}
\seealso{
refs_tiles
}
